Task8/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
wordcloud>=1.9.0
reportlab>=4.0.0
python-multipart>=0.0.6
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
matplotlib>=3.8.0
